requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
langchain_groq
streamlit
duckduckgo-search
langchain-groq
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
youtube-transcript-api
crewai
crewai_tools
load_dotenv
langchain-huggingface
langchain_openai
neo4j==5.14
langchain
langchain_community
langchain_openai
langchain_groq
python-dotenv
streamlit
langchain_experimental